Move continuations to ComponentContinuationManager
